+++ pages/paintings.tsx
@@ -1,12 +1,10 @@
-import Footer from "../src/Components/Footer";
-import Navbar from "../src/Components/Navbar";
+import React from "react";
+import PaintingsMain from "../src/Components/paintingsPage/PaintingMain";
 
 export default function Paintings() {
   return (
-    <div>
-      <Navbar />
-      {/* <Gallery /> */}
-      <Footer />
+    <div className="flex flex-col mt-[50px] md:mt-[0px]">
+      <PaintingsMain />
     </div>
   );
 }
